# backend/openapi.yaml
openapi: 3.1.0
info:
    title: Catelog API
    version: 1.0.0
servers:
    - url: /api
security:
    - bearerAuth: []
paths:
...
    /user/auth:
        get:
            summary: Start OAuth flow
            tags: [Auth]
            security: []
            parameters:
                - name: provider
                  in: query
                  required: true
                  schema:
                      type: string
                - name: redirect
                  in: query
                  required: true
                  schema:
                      type: string
            responses:
                302:
                    description: Redirect to OAuth
                400:
                    description: Invalid provider
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /user/auth/google/callback:
        get:
            summary: Google OAuth callback
            tags: [Auth]
            security: []
            parameters:
                - name: code
                  in: query
                  required: true
                  schema:
                      type: string
                - name: state
                  in: query
                  required: true
                  schema:
                      type: string
            responses:
                302:
                    description: Redirect to frontend with auth result
                400:
                    description: Invalid OAuth session
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /user/auth/microsoft/callback:
        get:
            summary: Microsoft OAuth callback
            tags: [Auth]
            security: []
            parameters:
                - name: code
                  in: query
                  required: true
                  schema:
                      type: string
                - name: state
                  in: query
                  required: true
                  schema:
                      type: string
            responses:
                302:
                    description: Redirect to frontend with auth result
                400:
                    description: Invalid OAuth session
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /user/auth/refresh:
        post:
            summary: Exchange a refresh token for a new access token
            tags: [Auth]
            security: []
            responses:
                200:
                    description: New token issued
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [accessToken]
                                properties:
                                    accessToken:
                                        type: string
                401: { $ref: "#/components/responses/UnauthorizedError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /user/auth/logout:
        post:
            summary: Revoke the refresh token and clear cookies
            tags: [Auth]
            security: []
            responses:
                200:
                    description: Successfully logged out
                500: { $ref: "#/components/responses/InternalServerError" }
    /user/me:
        get:
            summary: Get current authenticated user profile
            tags: [User]
            responses:
                200:
                    description: The authenticated user object
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/User" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
        patch:
            summary: Update user profile
            tags: [User]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                email:
                                    type: string
                                notificationsEnabled:
                                    type: boolean
            responses:
                200:
                    description: Profile updated
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/User" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
        delete:
            summary: Delete user account and all connected data
            tags: [User]
            responses:
                200:
                    description: Account deleted successfully
                401: { $ref: "#/components/responses/UnauthorizedError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /user/devices/subscribe:
        post:
            summary: "Register a device for push notifications"
            tags: [User]
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/UserDevice" }
            responses:
                200:
                    description: "Device subscribed successfully"
                401: { $ref: "#/components/responses/UnauthorizedError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /user/devices/unsubscribe:
        post:
            summary: "Unsubscribe a device from push notifications"
            tags: [User]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [endpoint]
                            properties:
                                endpoint:
                                    type: string
            responses:
                200:
                    description: "Device unsubscribed successfully"
                401: { $ref: "#/components/responses/UnauthorizedError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /titles:
        post:
            summary: Add a title to catalogue
            tags: [Titles]
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/Title" }
            responses:
                200:
                    description: Title added
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Title" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /titles/{id}:
        get:
            summary: Get title by ID
            tags: [Titles]
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                200:
                    description: Title details
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Title" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /titles/{id}/refresh:
        post:
            summary: Refresh title metadata or discover candidates
            tags: [Titles]
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                200:
                    description: Refresh triggered successfully, returns the updated title state
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Title" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /titles/search:
        get:
            summary: Search for a title in external sources
            tags: [Titles]
            parameters:
                - name: q
                  in: query
                  description: Search query (title name)
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    description: List of matching titles
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/Title" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                500: { $ref: "#/components/responses/InternalServerError" }
...
    /invites:
        get:
            summary: Get all invites related to the current user
            tags: [Invites]
            parameters:
                - name: t
                  in: query
                  schema:
                      type: string
                      enum: [incoming, outgoing]
            responses:
                200:
                    description: List of invites
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/Invite" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                500: { $ref: "#/components/responses/InternalServerError" }
        post:
            summary: Create a new invite
            tags: [Invites]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [listId, invitee]
                            properties:
                                listId:
                                    type: string
                                    format: uuid
                                invitee:
                                    type: string
                                    format: uuid
            responses:
                200:
                    description: Invite created
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /invites/{token}:
        get:
            summary: Get invite details by token
            tags: [Invites]
            security: []
            parameters:
                - name: token
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    description: Invite metadata for display
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Invite" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /invites/{token}/accept:
        post:
            summary: Accept watchlist invite
            tags: [Invites]
            parameters:
                - name: token
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    description: Invite accepted
                400:
                    description: Invite expired
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
    /invites/{id}/decline:
        delete:
            summary: Decline watchlist invite
            tags: [Invites]
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                200:
                    description: Invite declined
                401: { $ref: "#/components/responses/UnauthorizedError" }
                403: { $ref: "#/components/responses/ForbiddenError" }
                404: { $ref: "#/components/responses/NotFoundError" }
                500: { $ref: "#/components/responses/InternalServerError" }
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
        systemKey:
            type: apiKey
            in: header
            name: X-System-Key
    responses:
...
        UnauthorizedError: # 401
            description: Access token is missing or invalid
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        ForbiddenError: # 403
            description: The authenticated user does not have permission to access this resource
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        NotFoundError: # 404
            description: The requested resource was not found
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
        InternalServerError: # 500
            description: There was an unexpected error
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
    schemas:
        Error:
            type: object
            properties:
                message:
                    type: string
                errors: # For 400 Validation errors
                    type: array
                    items:
                        type: object
        Title:
            type: object
            required: [id, type, title, public, externalIds]
            readOnly: [id, createdAt, updatedAt]
            properties:
                id:
                    type: string
                    format: uuid
                type:
                    type: string
                    enum: [movie, series]
                title:
                    type: string
                localizedTitles:
                    type: object
                    additionalProperties:
                        type: string
                poster:
                    type: string
                year:
                    type: integer
                genres:
                    type: array
                    items:
                        type: string
                        enum:
                            [
                                action,
                                adventure,
                                comedy,
                                drama,
                                fantasy,
                                horror,
                                history,
                                sci_fi,
                                romance,
                                thriller,
                                animation,
                                documentary,
                                crime,
                                mystery,
                                family,
                                musical,
                            ]
                ratings:
                    type: object
                    additionalProperties:
                        type: number
                avgRating:
                    type: number
                directors:
                    type: array
                    items:
                        type: string
                actors:
                    type: array
                    items:
                        type: string
                durationMinutes:
                    type: integer
                externalIds:
                    type: object
                    additionalProperties:
                        type: string
                mergeCandidates:
                    type: array
                    items: { $ref: "#/components/schemas/MergeCandidate" }
                public:
                    type: boolean
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
        MergeCandidate:
            type: object
            required: [origin, displayData]
            properties:
                origin:
                    type: string
                    enum: [internal, tmdb, imdb, csfd]
                internalId:
                    type: string
                    format: uuid
                externalId:
                    type: string
                displayData:
                    type: object
                    required: [title]
                    properties:
                        title:
                            type: string
                        year:
                            type: integer
                        poster:
                            type: string
...
        User:
            type: object
            required: [id, email, createdAt]
            readOnly: [id, createdAt, updatedAt]
            properties:
                id:
                    type: string
                    format: uuid
                email:
                    type: string
                    format: email
                name:
                    type: string
                notificationsEnabled:
                    type: boolean
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
        UserDevice:
            type: object
            required: [endpoint, keys]
            readOnly: [id, createdAt]
            properties:
                id:
                    type: string
                    format: uuid
                userId:
                    type: string
                    format: uuid
                deviceName:
                    type: string
                endpoint:
                    type: string
                keys:
                    type: object
                    required: [p256dh, auth]
                    properties:
                        p256dh:
                            type: string
                        auth:
                            type: string
                createdAt:
                    type: string
                    format: date-time
        Invite:
            type: object
            required: [id, listId, inviter, invitee, token, expiresAt, createdAt]
            readOnly: [id, createdAt]
            properties:
                id:
                    type: string
                    format: uuid
                listId:
                    type: string
                    format: uuid
                listName:
                    type: string
                inviter:
                    type: string
                    format: uuid
                inviterName:
                    type: string
                invitee:
                    type: string
                    format: uuid
                token:
                    type: string
                expiresAt:
                    type: string
                    format: date-time
                createdAt:
                    type: string
                    format: date-time