# frontend-pwa/src/api/models/Error.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Error = {
    message?: string;
    errors?: Array<Record<string, any>>;
};