# frontend-pwa/src/api/models/User.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type User = {
    id: string;
    email: string;
    name?: string;
    notificationsEnabled?: boolean;
    createdAt: string;
    updatedAt?: string;
};